[trainer] fix: use FullyAsyncLLMServerClient for async trainer
